Populated map with pins where the rat sightings are based off of the latitude and longitude of each rat sighting

+++ app/src/main/java/edu/gatech/cats/cats_2340/controllers/RatSightingScreen.java
@@ -66,8 +66,8 @@ public class RatSightingScreen extends AppCompatActivity {
         address_text.setText(rs.getAddr());
         city_text.setText(rs.getCity());
         borough_text.setText(rs.getBorough().toString());
-        latitude_text.setText(Float.toString(rs.getLat()));
-        longitude_text.setText(Float.toString(rs.getLong()));
+        latitude_text.setText(Double.toString(rs.getLat()));
+        longitude_text.setText(Double.toString(rs.getLong()));
 
     }
 }
